feat: Change algolia function to gatsby plugin (#3393)
* fix: lens version 339

* feat: Change algolia function to gatsby plugin

* fix: package

* feat: linkable search results

* changes

* changes
