+++ package.json
@@ -11,12 +11,11 @@
     "@prisma/client": "2.29.1",
     "@prisma/lens": "0.0.339",
     "@reach/router": "1.3.4",
-    "algoliasearch": "4.11.0",
+    "algoliasearch": "^4.13.1",
     "babel-plugin-styled-components": "1.13.3",
     "feather-icons": "^4.29.0",
     "gatsby": "2.32.13",
     "gatsby-image": "2.11.0",
-    "gatsby-plugin-algolia": "0.5.0",
     "gatsby-plugin-catch-links": "2.10.0",
     "gatsby-plugin-google-tagmanager": "2.11.0",
     "gatsby-plugin-layout": "1.10.0",
@@ -87,11 +86,12 @@
   "license": "Apache-2.0",
   "scripts": {
     "clean": "gatsby clean",
-    "build": "gatsby build --prefix-paths --log-pages",
+    "build": "npm run install-algolia && gatsby build --prefix-paths --log-pages",
     "dev": "gatsby develop",
     "serve": "npm run build && gatsby serve --prefix-paths -o",
     "format": "prettier . --write --check",
     "test": "jest",
+    "install-algolia": "cd ./plugins/gatsby-algolia-indexer && npm install",
     "postinstall": "husky install"
   },
   "devDependencies": {
