+++ plugins/gatsby-algolia-indexer/gatsby-node.js
@@ -1,0 +1,1 @@
+//noop
