+++ plugins/gatsby-algolia-indexer/index.js
@@ -1,0 +1,1 @@
+//noop
