+++ plugins/gatsby-algolia-indexer/mdx-to-searchable.js
@@ -1,0 +1,8 @@
+const remark = require('remark')
+const mdx = require('remark-mdx')
+const searchable = require('./remark-mdx-searchable')
+
+module.exports = (doc) => {
+  const result = remark().use(mdx).use(searchable).processSync(doc)
+  return result.data
+}
