+++ plugins/gatsby-algolia-indexer/options.js
@@ -1,0 +1,8 @@
+module.exports = (options) => ({
+  appId: null,
+  adminKey: null,
+  searchKey: null,
+  indexName: null,
+  query: '', // optional: use this to overwrite the default query.
+  ...options,
+})
