+++ plugins/gatsby-algolia-indexer/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "gatsby-algolia-indexer",
+  "version": "1.0.0",
+  "description": "Algolia",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "algoliasearch": "4.11.0",
+    "gatsby-plugin-algolia": "0.5.0",
+    "remark": "^12.0.1",
+    "remark-mdx": "^1.6.16"
+  }
+}
